python-script/requirements.txt
google-genai==1.27.0
pypdf==6.1.0
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
python-dotenv